Change depth's type in Node to remove cast
This is done in order to remove unnecessary casts.

+++ src/algorithm/mod.rs
@@ -50,7 +50,7 @@ impl Algorithm {
 
     /// Set the initial Node to a new state using the provided Goban.
     pub fn update_initial_state(&mut self, initial_state: Goban, last_move: BitBoard, player_captures: u8, opponent_captures:u8) {
-        let new_initial_node = Node::new(initial_state, crate::DEPTH as usize, last_move, false, player_captures, opponent_captures);
+        let new_initial_node = Node::new(initial_state, crate::DEPTH, last_move, false, player_captures, opponent_captures);
         self.initial = new_initial_node;
     }
